modify: 차트 메모리 정리
