+++ src/components/ProfitLossChart.tsx
@@ -1,128 +1,22 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useRef } from "react";
 import Chart from "chart.js/auto";
-import { apiClient } from "@/lib/apiClient";
-import { useMarketStore } from "@/store/marketStore";
-import { useAssetStore } from "@/store/assetStore";
-
-type TradeHistory = {
-  concludedAt: string;
-  marketCode: string;
-  orderPosition: "BUY" | "SELL";
-  orderType: string;
-  tradePrice: number;
-  tradeQuantity: number;
-};
 
 type DataByDate = {
   date: string; // YYYY-MM-DD
-  cumulativeProfitLoss: number; // 누적 손익(원)
+  profitLoss: number; // 누적 손익(원)
+  profitLossRate: number;
 };
 
-export default function PortfolioChart() {
+interface Props {
+  chartData: DataByDate[];
+  isTradeHistoryLoading: boolean;
+  chartError: string | null;
+}
+export default function PortfolioChart({chartData, isTradeHistoryLoading, chartError}: Props) {
   const chartRef = useRef<Chart | null>(null);
   const canvasRef = useRef<HTMLCanvasElement | null>(null);
-  const { tradeHistory, fetchTradeHistory, isTradeHistoryLoading } = useAssetStore();
-  const [chartData, setChartData] = useState<DataByDate[]>([]);
-  const [error, setError] = useState<string | null>(null);
 
-  // Store 가져오기
-  const tickers = useMarketStore(state => state.tickers);
-  const assets = useAssetStore(state => state.assets);
-  const { getPeriodProfitLoss } = useAssetStore();
 
-  // 거래 내역 기반 누적 손익 계산 함수
-  const calculateCumulativeProfitLossByDate = (trades: TradeHistory[]) => {
-    if (!trades || trades.length === 0) {
-      console.log("거래 내역이 없습니다.");
-      return [];
-    }
-
-    if (!tickers || Object.keys(tickers).length === 0) {
-      console.log("티커 데이터가 로드되지 않았습니다.");
-      return [];
-    }
-
-    console.log(`거래 내역 ${trades.length}개 처리 중`);
-
-    // 날짜별로 거래 분류
-    const tradesByDate: Record<string, TradeHistory[]> = {};
-
-    // 날짜별로 거래 정렬 (가장 오래된 거래부터)
-    const sortedTrades = [...trades].sort((a, b) =>
-      new Date(a.concludedAt).getTime() - new Date(b.concludedAt).getTime()
-    );
-
-    // 각 거래의 날짜를 추출
-    sortedTrades.forEach(trade => {
-      const date = trade.concludedAt.slice(0, 10);
-      if (!tradesByDate[date]) {
-        tradesByDate[date] = [];
-      }
-      tradesByDate[date].push(trade);
-    });
-
-    // 날짜를 정렬
-    const sortedDates = Object.keys(tradesByDate).sort();
-
-    if (sortedDates.length === 0) {
-      console.log("정렬된 날짜가 없습니다.");
-      return [];
-    }
-
-    // 누적 거래 내역
-    let accumulatedTrades: TradeHistory[] = [];
-    const dataByDate: DataByDate[] = [];
-
-    // 각 날짜별로 누적 손익 계산
-    sortedDates.forEach(date => {
-      // 이 날짜까지의 모든 거래 누적
-      accumulatedTrades = [...accumulatedTrades, ...tradesByDate[date]];
-
-      try {
-        // assetStore의 메소드를 사용해 누적 손익 계산
-        const profitLossData = getPeriodProfitLoss(accumulatedTrades, tickers, 30);
-
-        // 결과 저장 (수익률 대신 손익액 사용)
-        dataByDate.push({
-          date,
-          cumulativeProfitLoss: Math.floor(profitLossData.periodProfitLoss) // 소수점 버림
-        });
-      } catch (err) {
-        console.error(`날짜 ${date} 처리 중 오류 발생:`, err);
-      }
-    });
-
-    console.log(`계산된 데이터 ${dataByDate.length}개 생성`);
-    return dataByDate;
-  };
-
-  // 거래 내역 가져오기
-  useEffect(() => {
-    fetchTradeHistory();
-  }, [fetchTradeHistory]);
-
-  // 거래 내역으로부터 차트 데이터 계산
-  useEffect(() => {
-    if (isTradeHistoryLoading || !tradeHistory.length || !tickers || Object.keys(tickers).length === 0) {
-      return;
-    }
-
-    try {
-      console.log("차트 데이터 계산 중...");
-      const data = calculateCumulativeProfitLossByDate(tradeHistory);
-      console.log("계산된 차트 데이터:", data);
-
-      if (data.length > 0) {
-        setChartData(data);
-        setError(null);
-      } else {
-        setError("계산된 데이터가 없습니다.");
-      }
-    } catch (err) {
-      console.error("차트 데이터 계산 중 오류:", err);
-      setError("데이터 계산 중 오류가 발생했습니다.");
-    }
-  }, [tradeHistory, tickers, isTradeHistoryLoading]);
 
   // 차트 그리기 (단 한번만 그리도록 설정)
   useEffect(() => {
@@ -136,7 +30,7 @@ export default function PortfolioChart() {
       if (!ctx) return;
 
       // 데이터 값 얻기
-      const values = chartData.map(item => item.cumulativeProfitLoss);
+      const values = chartData.map(item => item.profitLoss);
       const lastValue = values[values.length - 1] || 0;
 
       // 최대, 최소값 찾기 (적절한 Y축 범위 설정을 위해)
@@ -248,6 +142,11 @@ export default function PortfolioChart() {
     };
 
     drawChart();
+
+    return () => {
+      chartRef.current?.destroy();
+      chartRef.current = null;
+    }
   }, [chartData]);
 
   return (
@@ -257,9 +156,9 @@ export default function PortfolioChart() {
           <div className="absolute inset-0 flex justify-center items-center">
             <div className="animate-spin rounded-full h-12 w-12 border-4 border-blue-500 border-t-transparent" />
           </div>
-        ) : error ? (
+        ) : chartError ? (
           <div className="absolute inset-0 flex justify-center items-center text-gray-500">
-            {error}
+            {chartError}
           </div>
         ) : chartData.length === 0 ? (
           <div className="absolute inset-0 flex justify-center items-center text-gray-500">
