+++ src/lib/chartUtils.tsx
@@ -29,20 +29,19 @@ type TimeUnit = 'millisecond' | 'second' | 'minute' | 'hour' | 'day' | 'week' | 
 interface WriteChartProps {
     market: string;
     candle: Candle[];
-    canvasRef: React.RefObject<HTMLCanvasElement>;
     timeUnit: TimeUnit;
 }
 
 // candlestick 차트
-const WriteChart: React.FC<WriteChartProps> = ({ market, candle, canvasRef, timeUnit }) => {
+const WriteChart: React.FC<WriteChartProps> = ({ market, candle, timeUnit }) => {
 
      let chart_data: {x: number, y: number, data: any} | null = null;
      const chartRef = useRef<Chart | null>(null);
-     const animationFrameRef = useRef<number | null>(null);
+     const canvasRef = useRef<HTMLCanvasElement>(null);
     
     useEffect(() => {
         // 캔버스 요소 가져오기
-        const ctx = canvasRef.current?.getContext("2d");
+        const ctx = canvasRef.current?.getContext("2d")
         
         // 캔버스가 존재하지 않으면 종료
         if (!ctx) return;
@@ -78,11 +77,6 @@ const WriteChart: React.FC<WriteChartProps> = ({ market, candle, canvasRef, time
         if (chartRef.current) {
             chartRef.current.destroy();
             chartRef.current = null;
-        }
-
-        // 애니메이션 프레임 정리
-        if (animationFrameRef.current) {
-            cancelAnimationFrame(animationFrameRef.current);
         }
 
         // 차트 생성
@@ -161,119 +155,20 @@ const WriteChart: React.FC<WriteChartProps> = ({ market, candle, canvasRef, time
                         },
                     },
                 },
-                // 마우스가 차트 위에 있을 때 이벤트 발생
-                // onHover: function (event, chartElement, chart) {
-
-                //     if (event.native) {
-                //         // 차트 위 가장 가까운 차트 요소 가지고 옴
-                //         const elements = chart.getElementsAtEventForMode(
-                //             event.native,
-                //             "nearest",
-                //             { intersect: false },
-                //             true
-                //         );
-                //         // 데이터를 lastCrosshair에 저장
-                //         if (elements.length) {
-                //             const index = elements[0].index;
-                //             const data = chart.data.datasets[0].data[index];
-                //             if(event && event.x !== null && event.y !== null) {
-                //                 chart_data = {x: event.x, y: event.y, data: data};
-                //             }
-                //         }
-                //     }
-                // },
             },
     });
 
     chartRef.current = chart;
 
-    // if(chart.ctx) {
-    //       drawCrosshair();
-    // }
-
-    // // 교차선 구현
-    // function drawCrosshair() {
-    //     const chart = chartRef.current;
-    //     if (!chart) return;
-
-    //     const chartArea = chart.chartArea;
-    //     const ctx = chart.ctx;
-    //     const xAxis = chart.scales['x'];
-    //     const yAxis = chart.scales['y'];
-
-        
-    //     try {
-    //         chart.update('none'); // 애니메이션 없이 업데이트
-    //     } catch (e: any) {
-    //         if (e.message.includes("Cannot read properties of null (reading 'ownerDocument')")) {
-    //             // 에러 무시
-    //         } else {
-    //             return e;
-    //         }
-    //     }
-
-                
-    //     if (chart_data) {
-    //         const {x: X, y, data} = chart_data;
-    //         const {o, h, l, c, x: time} = data;
-
-    //         // cavas 요소가 로드 안된 경우
-    //         if(!ctx) {
-    //             return;
-    //         }
-                
-    //         ctx.save();
-    //         ctx.beginPath();
-    //         ctx.setLineDash([]); // 실선
-    //         ctx.strokeStyle = "#0000000" // 선 색상
-    //         ctx.lineWidth = 1; // 선 굵기
-
-    //         ctx.fillStyle = 'black'; // 글씨
-    //         ctx.font = '8px Arial'; // 폰트 설정
-    //         ctx.textAlign = 'center';
-    //         ctx.textBaseline = 'top' // 텍스트 위치
-                
-    //         // 수직선
-    //         ctx.moveTo(X, chartArea.top);
-    //         ctx.lineTo(X, chartArea.bottom);
-                
-    //         // 수평선 
-    //         ctx.moveTo(chartArea.left, y);
-    //         ctx.lineTo(chartArea.right, y);
-
-    //         // 데이터 정보
-    //         ctx.fillText(`시가: ${o}`, xAxis.left + 32, yAxis.top + 10);
-    //         ctx.fillText(`고가: ${h}`, xAxis.left + 32, yAxis.top + 20);
-    //         ctx.fillText(`시간: ${format(time, 'yyyy.MM.dd HH:mm:ss')}`, xAxis.left + 48, yAxis.top + 30);
-    //         ctx.fillText(`저가: ${l}`, xAxis.left + 110, yAxis.top + 10);
-    //         ctx.fillText(`종가: ${c}`, xAxis.left + 110, yAxis.top + 20);
-                
-    //         ctx.stroke();
-    //         ctx.restore();
-    //     }
-        
-    //     // 성능 최적화: 마우스가 차트 위에 있을 때만 교차선 그리기
-    //     if (chart_data) {
-    //         animationFrameRef.current = requestAnimationFrame(drawCrosshair);
-    //     } else {
-    //         animationFrameRef.current = requestAnimationFrame(drawCrosshair);
-    //     }
-    // };
-
     // 컴포넌트 언마운트 시 정리
     return () => {
-        if (animationFrameRef.current) {
-            cancelAnimationFrame(animationFrameRef.current);
-        }
-        if (chartRef.current) {
-            chartRef.current.destroy();
-            chartRef.current = null;
-        }
+        chartRef.current?.destroy();
+        chartRef.current = null;
     };
 
 }, [market, candle, timeUnit]);
 
-    return null;
+    return <canvas ref={canvasRef} />
 };
 
 export default WriteChart;
